+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+* Fix issue where `off` would throw an error if no events existed for named event
+
 # 0.2.0
 
 * Add `off` method